+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Add `Plugin` class and register package as a composer plugin
+- Add `Symlinker` class
 
 ### Removed
 - Remove support for Composer 1